+++ packages/api/src/controllers/MainController.js
@@ -10,8 +10,7 @@ const { base58 } = require('@scure/base')
 const DashCoreRPC = require('../dashcoreRpc')
 const TokensDAO = require('../dao/TokensDAO')
 const PlatformAddressesDAO = require('../dao/PlatformAddressesDAO')
-const { getPlatformQuorums, buildProposerSchedule } = require('../utils')
-const PaginatedResultSet = require('../models/PaginatedResultSet')
+const { getPlatformQuorums } = require('../utils')
 
 const API_VERSION = require('../../package.json').version
 
@@ -289,46 +288,6 @@ class MainController {
     }
 
     response.send(currentQuorum)
-  }
-
-  getProposers = async (request, response) => {
-    const { page = 1, limit = 10 } = request.query
-
-    const [{ quorumHash: currentQuorumHash }, { highestBlock }] = await Promise.all([
-      TenderdashRPC.getValidators(),
-      TenderdashRPC.getStatus()
-    ])
-
-    const lastBlock = await TenderdashRPC.getBlockByHeight(highestBlock.height)
-
-    const lastProposer = lastBlock?.block?.header?.proposer_pro_tx_hash?.toUpperCase() ?? null
-
-    const { quorums } = await getPlatformQuorums()
-
-    // the schedule ends when the rotation has been walked once, so building it whole gives
-    // both the total and the page without repeating the walk
-    const schedule = buildProposerSchedule(
-      quorums,
-      currentQuorumHash?.toUpperCase() ?? null,
-      lastProposer,
-      highestBlock.height,
-      quorums.reduce((total, quorum) => total + (quorum.members ?? []).length, 0)
-    )
-
-    if (!schedule) {
-      return response.status(404).send({ message: 'not found' })
-    }
-
-    const fromRank = (Number(page) - 1) * Number(limit)
-
-    response.send(
-      new PaginatedResultSet(
-        schedule.slice(fromRank, fromRank + Number(limit)),
-        Number(page),
-        Number(limit),
-        schedule.length
-      )
-    )
   }
 
   getQuorumByHash = async (request, response) => {
